requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
tensorflow>=2.16.1
tensorflow-hub>=0.16.1
Pillow>=10.1.0
numpy>=1.26.2
opencv-python>=4.8.1.78
waitress==2.1.2
